Develop

+++ src/api/routes/user_authentication.py
@@ -1,6 +1,11 @@
-from fastapi import APIRouter
+from fastapi import APIRouter, Depends, HTTPException
 from fastapi import File, UploadFile, Form
 from src.schemas.registration import Register
+from src.utils.password_utils import Hash
+from sqlalchemy.orm import Session
+from src.models.model import Candidates
+from src.utils.email_utils import GenerateMail
+from db import get_db
 import json
 import os
 
@@ -10,15 +15,43 @@ router = APIRouter(prefix="/auth")
 UPLOAD_DIR = os.path.join("src", "files")
 
 @router.post('/register')
-async def candidate_registration(payload: str = Form(...),file: UploadFile = File(...)):
+async def candidate_registration(payload: str = Form(...),file: UploadFile = File(...), db: Session = Depends(get_db)):
     try:
-        data = Register(**json.loads(payload))
-        file_location = os.path.join(UPLOAD_DIR, file.filename)
+        data_dict = json.loads(payload)
+        data = Register(**data_dict)
 
+        existing_user = db.query(Candidates).filter(Candidates.email == data.email).first()
+        if existing_user:
+            raise HTTPException(status_code=400, detail="User already exists")
+        
+
+        encrypt_password = Hash.get_password_hash(data.password)
+
+        file_location = os.path.join(UPLOAD_DIR, file.filename)
         with open(file_location, "wb") as f:
             f.write(await file.read())
-
-        return {"message": "Registered successfully", "file_saved_at": file_location,"name":data.username}
+        print(file.filename)
+       
+        new_user = Candidates(
+            role_id = 3,
+            username=data.username,
+            email=data.email,
+            contact=data.contact,
+            dob=data.dob,
+            gender = data.gender,
+            github_url = data.github,
+            linkedin_url = data.linkedin,
+            resume = file.filename,
+            has_experience = data.experienced,
+            password= encrypt_password
+        )
+        email = data.email
+        GenerateMail.send_email(email)
+        db.add(new_user)
+        db.commit()
+        db.refresh(new_user)
+     
+        return {"message": "Registered successfully"}
 
     except Exception as  e:
         print(e)
